Split BuildContext<T>
