+++ src/AutoRest.CSharp/Output/Models/Types/BuildContext.cs
@@ -6,7 +6,6 @@ using AutoRest.CSharp.Generation.Types;
 using AutoRest.CSharp.Input;
 using AutoRest.CSharp.Input.Source;
 
-#pragma warning disable SA1402
 namespace AutoRest.CSharp.Output.Models.Types
 {
     internal class BuildContext
@@ -26,26 +25,5 @@ namespace AutoRest.CSharp.Output.Models.Types
         public Configuration Configuration { get; }
         public SourceInputModel? SourceInputModel { get; }
         public virtual TypeFactory TypeFactory { get; } = null!;
-    }
-
-    internal class BuildContext<T> : BuildContext where T: OutputLibrary
-    {
-        private TypeFactory? _typeFactory;
-
-        public T Library { get; private set; }
-
-        public BuildContext(CodeModel codeModel, Configuration configuration, SourceInputModel? sourceInputModel): base(codeModel, configuration, sourceInputModel)
-        {
-            if (configuration.LowLevelClient)
-            {
-                Library = (T)(object)new LowLevelOutputLibrary(codeModel, (BuildContext<LowLevelOutputLibrary>)(object)this);
-            }
-            else
-            {
-                Library = (T)(object)new DataPlaneOutputLibrary(codeModel, (BuildContext<DataPlaneOutputLibrary>)(object)this);
-            }
-        }
-
-        public override TypeFactory TypeFactory => _typeFactory ??= new TypeFactory(Library);
     }
 }
